+++ app/Actions/Import/Exec.php
@@ -5,9 +5,9 @@ namespace App\Actions\Import;
 use App\Actions\Album\Create as AlbumCreate;
 use App\Actions\Photo\Create as PhotoCreate;
 use App\Actions\Photo\Strategies\ImportMode;
-use App\DTO\BaseImportReport;
-use App\DTO\ImportEventReport;
-use App\DTO\ImportProgressReport;
+use App\Data\Reports\BaseImportReport;
+use App\Data\Reports\ImportEventReport;
+use App\Data\Reports\ImportProgressReport;
 use App\Exceptions\FileOperationException;
 use App\Exceptions\Handler;
 use App\Exceptions\ImportCancelledException;
