+++ app/Data/Reports/ImportEventReport.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\DTO;
+namespace App\Data\Reports;
 
 use App\Enum\SeverityType;
 use App\Exceptions\Handler as ExceptionHandler;
@@ -9,20 +9,14 @@ class ImportEventReport extends BaseImportReport
 {
 	public const REPORT_TYPE = 'event';
 
-	protected string $subtype;
-	protected ?string $path;
-	protected SeverityType $severity;
-	protected string $message;
-	protected ?\Throwable $throwable;
-
-	protected function __construct(string $subtype, SeverityType $severity, ?string $path, string $message, ?\Throwable $throwable = null)
+	protected function __construct(
+		public string $subtype,
+		public SeverityType $severity,
+		public ?string $path,
+		public string $message,
+		protected ?\Throwable $throwable = null)
 	{
 		parent::__construct(self::REPORT_TYPE);
-		$this->subtype = $subtype;
-		$this->severity = $severity;
-		$this->path = $path;
-		$this->message = $message;
-		$this->throwable = $throwable;
 	}
 
 	public static function createWarning(string $subtype, ?string $path, string $message): self
@@ -38,19 +32,6 @@ class ImportEventReport extends BaseImportReport
 	public function getException(): ?\Throwable
 	{
 		return $this->throwable;
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function toArray(): array
-	{
-		return array_merge(parent::toArray(), [
-			'subtype' => $this->subtype,
-			'severity' => $this->severity->value,
-			'path' => $this->path,
-			'message' => $this->message,
-		]);
 	}
 
 	public function toCLIString(): string
