+++ app/Exceptions/Handler.php
@@ -3,7 +3,7 @@
 namespace App\Exceptions;
 
 use App\Contracts\Exceptions\Handlers\HttpExceptionHandler;
-use App\DTO\BacktraceRecord;
+use App\Data\BacktraceRecord;
 use App\Enum\SeverityType;
 use App\Exceptions\Handlers\AccessDBDenied;
 use App\Exceptions\Handlers\AdminSetterHandler;
